add column status in vihicles and refactoring vehicles and employee

+++ src/services/Employee/DeleteEmployeeService.ts
@@ -5,6 +5,10 @@ class DeleteEmployeeService {
   async execute({ id }) {
     const employee = await employeeRepository.findOneBy({ id })
 
+    if (!employee) {
+      throw new Error('Funcionário não existe')
+    }
+
     await employeeRepository.delete(employee.id)
   }
 }
