+++ src/services/Vehicle/DeleteVehicleService.ts
@@ -5,6 +5,10 @@ class DeleteVehicleService {
   async execute({ chassi }) {
     const vehicle = await vehicleRepository.findOneBy({ chassi })
 
+    if (!vehicle) {
+      throw new Error('Veículo não cadastrado')
+    }
+
     await vehicleRepository.delete(vehicle.id)
   }
 }
